Fix bugs in prime sieves.

+++ Number-Theory/prime_sieve.cpp
@@ -4,16 +4,16 @@
 /* Sieve of eratosthenes */
 
 /* Returns a std::vector of size n*/
-std::vector<bool> bool_prime_sieve(int n){
+std::vector<bool> bool_prime_sieve(long long n){
     std::vector<bool> prime(n + 1, true);
     prime[0] = prime[1] = false;
-    for(int i = 4; i <= n ; i += 2){
+    for(long long i = 4; i <= n ; i += 2){
         prime[i] = false;
     }
 
-    for(int i = 3 ; i * i <= n ; i += 2){
+    for(long long i = 3 ; i * i <= n ; i += 2){
         if(prime[i]){
-            for(int j = i * i; j <= n ; j += i){
+            for(long long j = i * i; j <= n ; j += i){
                 prime[j] = false;
             }
         }
@@ -28,17 +28,17 @@ std::vector<bool> bool_prime_sieve(int n){
     same as what you define in the function.
     (Just for sanity)
 */
-std::bitset<100000> bits_prime_sieve(int n){
+std::bitset<100000> bits_prime_sieve(long long n){
     std::bitset<100000> prime;
     prime.set();
     prime[0] = prime[1] = 0;
-    for(int i = 4; i <= n ; i += 2){
+    for(long long i = 4; i <= n ; i += 2){
         prime[i] = 0;
     }
 
-    for(int i = 3 ; i * i <= n ; i += 2){
+    for(long long i = 3 ; i * i <= n ; i += 2){
         if(prime[i]){
-            for(int j = i * i; j <= n ; j += i){
+            for(long long j = i * i; j <= n ; j += i){
                 prime[j] = 0;
             }
         }
